backend/requirements.txt
fastapi
uvicorn
supabase
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
httpx
